lesson7

+++ core/src/ru/jevo/animation/screens/GameScreen.java
@@ -7,21 +7,29 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.MathUtils;
+
+import java.util.ArrayList;
+import java.util.List;
+
 import ru.jevo.animation.basic.BasicScreen;
 import ru.jevo.animation.basic.Ship;
+import ru.jevo.animation.basic.Sprite;
+import ru.jevo.animation.pools.other.ExplosionPool;
 import ru.jevo.animation.pools.ships.ColonyShipPool;
 import ru.jevo.animation.pools.ships.MilitaryLargePool;
 import ru.jevo.animation.pools.ships.MilitaryMediumPool;
 import ru.jevo.animation.pools.weapons.BulletPool;
 import ru.jevo.animation.pools.ships.MilitarySmallPool;
 import ru.jevo.animation.pools.weapons.SimpleBlasterPool;
-import ru.jevo.animation.sprites.ships.ColonyShip;
+import ru.jevo.animation.sprites.button_titles.GameOverTitle;
+import ru.jevo.animation.sprites.button_titles.NewGameTitle;
 import ru.jevo.animation.sprites.ships.MilitaryLarge;
 import ru.jevo.animation.sprites.ships.MilitaryMedium;
 import ru.jevo.animation.sprites.ships.MilitarySmall;
 import ru.jevo.animation.sprites.ships.Viking;
-import ru.jevo.animation.sprites.Background;
-import ru.jevo.animation.sprites.Star;
+import ru.jevo.animation.sprites.other.Background;
+import ru.jevo.animation.sprites.other.Star;
+import ru.jevo.animation.sprites.weapon.Bullet;
 
 
 public class GameScreen extends BasicScreen {
@@ -30,13 +38,16 @@ public class GameScreen extends BasicScreen {
     private float animateTime = 0;
 
     Viking mViking;
-    TextureAtlas mainTextureAtlas;
+    private NewGameTitle titleNewGame;
+    private GameOverTitle titleGameOver;
     private BulletPool bulletPool;
     private SimpleBlasterPool blasterPool;
     private MilitarySmallPool mMilitarySmallPool;
     private MilitaryMediumPool mMilitaryMediumPool;
     private MilitaryLargePool mMilitaryLargePool;
     private ColonyShipPool mColonyShipPool;
+    private ExplosionPool mExplosionPool;
+    private List<Ship> ships = new ArrayList<Ship>();
 
     public GameScreen(Game game) {
         super(game);
@@ -46,11 +57,12 @@ public class GameScreen extends BasicScreen {
     public void show() {
         super.show();
         bgTexture = new Texture("images/bg.png");
-        menuTextureAtlas = new TextureAtlas("atlas/menuAtlas.tpack");
-        mainTextureAtlas = new TextureAtlas("atlas/mainAtlas.tpack");
+        titleNewGame = new NewGameTitle(mainTextureAtlas);
+        titleGameOver = new GameOverTitle(mainTextureAtlas);
         mBackground = new Background(new TextureRegion(bgTexture));
         rainMusic.setLooping(true); // повторение музыки
         rainMusic.play();
+        mExplosionPool = new ExplosionPool(mainTextureAtlas);
         mStar = new Star[STAR_COUNT];
         for (int i = 0; i < STAR_COUNT; i++)
             mStar[i] = new Star(menuTextureAtlas);
@@ -66,17 +78,77 @@ public class GameScreen extends BasicScreen {
     @Override
     public void render(float delta) {
         super.render(delta);
-        update(delta);
-        checkCollisions();
-        deleteAllDestroyed();
-        draw(delta);
+        if (mViking.gethP() > 0) {
+            update(delta);
+            checkCollisions();
+            deleteAllDestroyed();
+            draw(delta);
+        } else {
+            titleGameOver.update(delta);
+            Gdx.gl.glClearColor(0, 0, 0, 1);
+            Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
+            batch.begin();
+            mBackground.draw(batch);
+            titleGameOver.draw(batch);
+            batch.end();
+        }
     }
 
     private void checkCollisions() {
+        collisionInit();
+        collisionTaran(mExplosionPool);
+        collisionBulletWithEnemy();
+        collisionBulletWithViking();
+
+
     }
+
+    private void collisionBulletWithViking() {
+        List<Bullet> bulletList =  bulletPool.getactivePool();
+        for (Bullet bullet : bulletList) {
+            if (bullet.isDestroyed() || bullet.getOwner() == mViking) {
+                continue;
+            }
+            if (mViking.isBulletCollision(bullet)) {
+                bullet.setDestroyed(true);
+                mViking.damage(bullet.getDamage(), mExplosionPool);
+            }
+        }
+    }
+
+    private void collisionBulletWithEnemy() {
+        List<Bullet> bulletList =  bulletPool.getactivePool();
+        for (Ship enemy : ships) {
+            if (enemy.isDestroyed()) {
+                continue;
+            }
+            for (Bullet bullet : bulletList) {
+                if (bullet.getOwner() != mViking || bullet.isDestroyed()) {
+                    continue;
+                }
+                if (enemy.isBulletCollision(bullet)) {
+                    enemy.damage(bullet.getDamage(), mExplosionPool);
+                    bullet.setDestroyed(true);
+                }
+            }
+        }
+
+    }
+
+    private void collisionInit() {
+        List<MilitarySmall> minPool = mMilitarySmallPool.getactivePool();
+        List<MilitaryMedium> medPool = mMilitaryMediumPool.getactivePool();
+        List<MilitaryLarge> largePool = mMilitaryLargePool.getactivePool();
+        ships.addAll(minPool);
+        ships.addAll(medPool);
+        ships.addAll(largePool);
+    }
+
 
     public void update(float delta) {
         mViking.update(delta);
+        titleNewGame.update(delta);
+        mExplosionPool.updateActiveSprites(delta);
         bulletPool.updateActiveSprites(delta);
         mMilitarySmallPool.updateActiveSprites(delta);
         mMilitaryMediumPool.updateActiveSprites(delta);
@@ -91,6 +163,7 @@ public class GameScreen extends BasicScreen {
             animateTimer[0] += Math.sqrt(animateTimer[0]);
             MilitarySmall mMilitarySmallShip = mMilitarySmallPool.obtain();
             init(mMilitarySmallShip);
+
         }
         if (animateTimer[1] < animateTime) {
             animateTimer[1] += animateTimer[1];
@@ -112,6 +185,17 @@ public class GameScreen extends BasicScreen {
         }
         */
 
+
+    }
+
+    @Override
+    public void pause() {
+        System.out.println("Пауза");
+    }
+
+    @Override
+    public void resume() {
+        super.resume();
     }
 
     private void init(Ship ship) {
@@ -120,6 +204,7 @@ public class GameScreen extends BasicScreen {
     }
 
     public void deleteAllDestroyed() {
+        mExplosionPool.freeAllDestroyedActiveSprites();
         bulletPool.freeAllDestroyedActiveSprites();
         mMilitarySmallPool.freeAllDestroyedActiveSprites();
         mMilitaryMediumPool.freeAllDestroyedActiveSprites();
@@ -132,6 +217,7 @@ public class GameScreen extends BasicScreen {
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
         batch.begin();
         mBackground.draw(batch);
+        titleNewGame.draw(batch);
         for (int i = 0; i < STAR_COUNT; i++)
             mStar[i].draw(batch);
         mViking.draw(batch);
@@ -140,6 +226,7 @@ public class GameScreen extends BasicScreen {
         mMilitaryMediumPool.drawActiveSprites(batch);
         mMilitaryLargePool.drawActiveSprites(batch);
         mColonyShipPool.drawActiveSprites(batch);
+        mExplosionPool.drawActiveSprites(batch);
         batch.end();
     }
 
@@ -208,7 +295,23 @@ public class GameScreen extends BasicScreen {
         mColonyShipPool.dispose();
         mMilitaryLargePool.dispose();
         mMilitaryMediumPool.dispose();
+        mExplosionPool.dispose();
         super.dispose();
+    }
+
+    private void collisionTaran(ExplosionPool exp) {
+        for (Ship enemy : ships) {
+            if (enemy.isDestroyed())
+                continue;
+
+            float minDist = enemy.getHalfWidth() + mViking.getHalfWidth();
+            if (enemy.pos.dst2(mViking.pos) < minDist * minDist) {
+                enemy.setDestroyed(true);
+                enemy.boom(mExplosionPool);
+                mViking.damage(Viking.DAMAGE_TARAN,exp);
+                return;
+            }
+        }
     }
 
 }
